+++ src/canister_types/src/models/reward.rs
@@ -44,7 +44,7 @@ impl From<RewardableActivity> for RewardableActivityResponse {
     }
 }
 
-#[derive(CandidType, Deserialize, Serialize)]
+#[derive(CandidType, Deserialize, Serialize, Debug)]
 pub struct RewardableActivityResponse {
     pub activity: Activity,
     pub timestamp: u64,
@@ -53,7 +53,6 @@ pub struct RewardableActivityResponse {
 #[derive(Clone, Debug, CandidType, Deserialize, Serialize)]
 pub enum Activity {
     GroupMemberCount(u64),
-    UserActivity(Principal),
     UserReferral(Principal),
     UserProfileFilled(Principal),
     FirstGroupJoined(Principal),
@@ -73,7 +72,6 @@ impl fmt::Display for Activity {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
             Activity::GroupMemberCount(_) => write!(f, "group_member_count"),
-            Activity::UserActivity(_) => write!(f, "user_activity"),
             Activity::UserReferral(_) => write!(f, "user_referral"),
             Activity::UserProfileFilled(_) => write!(f, "user_profile_filled"),
             Activity::FirstGroupJoined(_) => write!(f, "first_group_joined"),
@@ -113,7 +111,6 @@ impl UserActivity {
 #[derive(CandidType, Deserialize, Clone)]
 pub struct RewardDataPackage {
     pub group_member_counts: Vec<GroupReward>,
-    pub user_activity: Vec<UserActivity>,
     pub user_referrals: Vec<Principal>,
     pub filled_profiles: Vec<Principal>,
     pub first_group_joined: Vec<Principal>,